Try to delay fetching until <authorized-image> is visible
When using <iron-image>/<img> the browser knows when (and if!) to fetch the image
content, based on whether the image would be visible right now. As we're side-stepping
that magic we now end up with fetching images for all <authorized-image> elements on a
page without discrimination, which can lead to horribly performance.

Use a very basic heuristic to only fetch when the component is visible, or had been
visible once. This is using IntersectionObserver if the browser supports it, otherwise
it is assumed that the component is visible.