Avoid repeatedly fetching when an error occurs and a token is set
When an error occurs our fetchDataUrl will return a result with an empty data URL and the correct 'src'
value, which will never be identical to any current `_displayed`, triggering a Polymer property-change.
If a token is set for the component however that change will lead to a forced reload (dataUrl is empty
but a token is set), which we do not want.