Avoid races when the token changes quickly
Simplify the logic for protecting against races and changes, and make _displayed contain all parameters
that were used to fetch the content. This allows us to also understand when the token has changed in any
way, and only accept the content when everything is still sane.

While there: Rename and hide the 'displayed' helper method as '_fetched', which matches better the intent
and use.